+++ src/charset/latin/default.js
@@ -2,10 +2,10 @@ import { EncoderOptions } from "../../type.js";
 
 /** @type EncoderOptions */
 const options = {
-    normalize: function(str){
-        return str.toLowerCase();
-    },
-    //numeric: false,
+    normalize: true
+    // normalize: function(str){
+    //     return str.toLowerCase();
+    // },
     //dedupe: false
 };
 export default options;