+++ src/charset/latin/exact.js
@@ -2,8 +2,10 @@ import { EncoderOptions } from "../../type.js";
 
 /** @type EncoderOptions */
 const options = {
-    split: /\s+/,
     normalize: false,
+    numeric: false,
+    split: /\s+/
+    //normalize: false,
     //dedupe: false
 };
 export default options;